+++ tests/codegen/fcode/scripts/issue_177.py
@@ -1,6 +1,5 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring, missing-class-docstring
 # coding: utf-8
-# pylint: disable=unused-variable
 
 from numpy import zeros
 from numpy import ones
@@ -11,16 +10,15 @@ def test_1():
 
     y = zeros((5,7))
     yp = y
+    return xp[0], yp[0,0]
 
 
-#$ header class Coords(public)
-#$ header method __init__(Coords, double [:])
 class Coords:
-    def __init__(self, m):
+    def __init__(self : 'Coords', m : 'float[:]'):
         self.major_radius = m
 
 m = ones(5)
 coor = Coords(m)
 print(coor.major_radius)
 
-test_1()
+print(test_1())