+++ tests/pyccel/test_pyccel.py
@@ -863,7 +863,6 @@ def test_basic_header():
 
 #------------------------------------------------------------------------------
 @pytest.mark.parametrize( "test_file", ["scripts/classes/classes.py",
-                                        "scripts/classes/generic_methods.py",
                                         "scripts/classes/classes_1.py",
                                         ] )
 @pytest.mark.parametrize( 'language', (
@@ -885,6 +884,7 @@ def test_classes_f_only( test_file , language):
                                         "scripts/classes/classes_3.py",
                                         "scripts/classes/classes_1.py",
                                         "scripts/classes/classes.py",
+                                        "scripts/classes/generic_methods.py",
                                         ] )
 @pytest.mark.parametrize( 'language', (
         pytest.param("python", marks = pytest.mark.python),